pstore-test: Ensure do_patch executes after linux-luv:do_patch
We introduced a placeholder REBOOT_DIR in the pstore selftests as
part of commit 251961 "linux-yocto-efi-test: pstore selftest reboot
flag" which ultimately should be replaced by the location of the
directory which survives a reboot in case we opt for the pstore
reboot tests.

Thus, we must ensure that the kernel source code is downloaded, the
placeholder is added and only then update the placeholder with the
actual value from the pstore-test.bb recipe. It so happens that this
dependency is not strict, hence we sometimes observe that the
pstore-test:do_patch runs earlier than the kernel's do_patch task
and thus placeholder is not replaced with actual contents, and the
pstore -tests do not execute as expected.

Changes V1->V2:
Change from do_patch[depends] += "linux-luv:do_patch" to
do_patch[depends] += "virtual/kernel:do_patch"

Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>