Plymouth: Revert commit that uses device timeout
After updating to plymouth 0.9.4, we see that if we execute the reboot
tests (and reboot through kexec), the splash screen does not come up.
All we see is the penguins (indicating Linux boot) and a blinking
cursor. On doing a git bisect, commit 7e37d58 "device-manager: use
device timeout instead of udev coldplug monitor" is the offending patch.

For now revert this patch (not a clean revert), as we are not sure why
this patch causes this behavior.

Acked-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>